+++ src/react/components/Icon/FontAwesomeIcon.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import '../../stylesheets/icons.scss';
+
 const propTypes = {
   name: PropTypes.string.isRequired,
   color: PropTypes.string
@@ -8,7 +10,7 @@ const propTypes = {
 
 const FontAwesomeIcon = (props) => {
   if (props.color) {
-    return <i className={`fa fa-${props.name} cc-icon cc-icon-${props.color}`}></i>;
+    return <i className={`fa fa-${props.name} cc-react-icon cc-react-icon--${props.color}`}></i>;
   } else {
     return <i className={`fa fa-${props.name}`}></i>;
   }
