+++ src/react/components/Icon/Icon.js
@@ -1,7 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import FontAwesomeIcon from './FontAwesomeIcon';
-import { ICON_PRINT, ICON_GRADUATION, ICON_CERTIFICATE } from './IconTypes';
+import {
+  ICON_CERTIFICATE,
+  ICON_CHECKMARK,
+  ICON_EXCLAMATION,
+  ICON_GRADUATION,
+  ICON_PRINT,
+  ICON_TIMES_CIRCLE
+} from './IconTypes';
 
 const propTypes = {
   name: PropTypes.string.isRequired
@@ -9,6 +16,12 @@ const propTypes = {
 
 const Icon = (props) => {
   switch (props.name) {
+    case ICON_CHECKMARK:
+      return <FontAwesomeIcon name={props.name} color='green' />;
+    case ICON_EXCLAMATION:
+      return <FontAwesomeIcon name={props.name} color='red' />;
+    case ICON_TIMES_CIRCLE:
+      return <FontAwesomeIcon name={props.name} color='red' />;
     case ICON_PRINT:
       return <FontAwesomeIcon name={props.name} />;
     case ICON_GRADUATION:
