+++ src/react/components/Icon/IconTypes.js
@@ -1,0 +1,6 @@
+export const ICON_CERTIFICATE = 'certificate';
+export const ICON_CHECKMARK = 'check';
+export const ICON_EXCLAMATION = 'exclamation-circle';
+export const ICON_GRADUATION = 'graduation-cap';
+export const ICON_PRINT = 'print';
+export const ICON_TIMES_CIRCLE = 'times-circle';
