--- src/react/components/IconTypes.js
@@ -1,3 +1,0 @@
-export const ICON_PRINT = 'print';
-export const ICON_GRADUATION = 'graduation-cap';
-export const ICON_CERTIFICATE = 'certificate';
