--- src/react/components/base/icon/GreenCheckmark.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import '../../../stylesheets/icons.scss';
-
-const GreenCheckmark = () => {
-  return (
-    <i className="fa fa-check cc-react-icon-green cc-react-icon"></i>
-  );
-};
-
-export default GreenCheckmark;
