--- src/react/components/base/icon/RedExclamationCircle.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import '../../../stylesheets/icons.scss';
-
-const RedExclamationCircle = () => {
-  return (
-    <i className="fa fa-exclamation-circle cc-react-icon--red cc-react-icon"></i>
-  );
-};
-
-export default RedExclamationCircle;
