--- src/react/components/base/icon/RedTimesCircle.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import '../../../stylesheets/icons.scss';
-
-const RedTimesCircle = () => {
-  return (
-    <i className="fa fa-times-circle cc-react-icon--red cc-react-icon"></i>
-  );
-};
-
-export default RedTimesCircle;
