+++ src/react/stylesheets/_colors.scss
@@ -1,5 +1,6 @@
 // Shades of Blue
 $cc-color-air-force-blue: #61889e;
+$cc-color-berkeley-blue: #003262;
 $cc-color-blue-lagoon: #065a7d;
 $cc-color-curious-blue: #2683bc;
 $cc-color-lochmara: #1f7099;
