+++ src/react/stylesheets/icons.scss
@@ -10,7 +10,11 @@
 }
 
 // Modifiers
-.cc-react-icon-green {
+.cc-react-icon--blue {
+  color: $cc-color-berkeley-blue;
+}
+
+.cc-react-icon--green {
   color: $cc-color-salem;
 }
